c/health_monitor: distribute health report collection
Change the health monitor logic to distribute the health report
collection logic. Previously all the nodes queried the cluster health
from the `redpanda/controller/0` partition leader. This put additional
pressure on that node as it had to deal with serialization of node
reports.

Changed health report collection logic so that every node queries each
other to collect its health report statistics. This way the overhead
related with serialization and handling health report request is evenly
distributed among all the nodes in the cluster.

Signed-off-by: Michal Maslanka <michal@redpanda.com>