rpk cloud login: change MakeAuthCurrent to **config.RpkCloudAuth
When we made an auth current, we reordered the underlying auth slice
which is _not_ pointer based.

Any changes to the pointer fields _after_ reodering were changes to
an old slice that would no longer be saved.

Now, we MakeAuthCurrent with two pointer indirects --
The underlying pointer to the slice element is used for the invariant
check, did we find this auth.
We now update the pointer-to-the-pointer so that if we update any fields
_after_ MakeAuthCurrent, they're mirrored into the new slice.