kafka: avoid disconnecting on storage errors in fetch handler
The storage layer can thow a few uncaught exceptions that are
caught in the fetch handler. I.e,
 - cloud_storage::download_exception
 - seastar::gate_closed_exception

When the fetch handler catches these it will simply close the client
connection and log the exception it caught. This isn't the most optimal
behavior as the errors caught from the storage layer are retriable. And
any data read from other partitions in the fetch request is lost when
the connection is closed.

This commit changes the behavior to instead return a retriable error for
the partition the storage exception occurs in.