mvlog: add initial entry header and body types
Introduces a storage-layer abstraction that is orthogonal to record
batches. This abstraction will be used to comprise segments that house
record batches, or additional metadata that may be useful to be appended
to the segment (e.g. truncation records, term information).

This only introduces the entry type for record batches, which will now
also include the term.

A later commit will introduce a parser to read entries. At a high level,
the steps will be similar to how record batches are read today: reading
the entry header first, then reading the entry body once the header is
validated. To that end, the entry header is fixed size rather than a
serde::envelope and is expected to not change much, while the bodies are
expected to be serde::envelopes that are more easily evolved.