c/health_monitor: introduced columnar format for partition statuses
Introduced a new way of storing partition statuses in health monitor.
Health monitor gathers information about every partition that exists in
the cluster to provide other components with the information on
partition health, their sizes and other runtime specific metadata. With
large number of topics and partition in the system serde operations are
computationally expensive and the data structures used cause significant
memory pressure.

In order to address that and increase Redpadna capabilities in terms of
supported partition and topic numbers replaced the old way of storing
partition statuses with new optimized columnar format. In columnar
representation values of each field of the `partition_status` data
structure are stored in one `delta_for_column`. All the topic partition
statuses are flattened into one columnar store. This way there is only
few columns to deal with when serializing and deserializing the data
structure. An additional optimization is to keep an index of all
namespaces instead of having it materialized for each topic in the
report.

Signed-off-by: Michal Maslanka <michal@redpanda.com>