Write segments to local Parquet as they are uploaded to TS.
This is the first step towards outputting Parquet files to support
Iceberg. It adds a class, `arrow_writing_consumer`, that can be used
with `record_batch_reader::consume` to write a log to a Parquet file.

This change also adds a call to arrow_writing_consumer from
ntp_archiver_service so that any time we write a segment to tiered
storage, we also output a local Parquet file.