mvlog: add class to collect record batches
Adds an abstraction that will be used on the read path to collect record
batches. This abstraction encapsulates some of the behavior that exists
in storage::log_segment_batch_reader and storage::skipping_consumer in
that its role is to determine whether it should collect a given record
as a part of a read, and then collect it.

Some later changes will introduce a new entry abstraction that will wrap
the entire record batch header and body to be fed into this collector.

I considered implementing the existing batch_consumer interface and
encapsulating some bits of the segment batch reader into a new reader
class, but felt like the batch_consumer overcomplicated the business
logic of batch collection.