crypto: Added short circuit to service startup
Cluster fixture tests appear to start multiple Redpanda instances
on the same thread.  This was resulting in an assertion during
service shutdown because the replaced context was not the expected
one (it was replaced by startup of the service on the same thread).

A new environment variable has been added to the CTest environment
when running a fixture test.  The short circuit now detects if the
global default context is the current context and if it is not,
will exit the start up of the service if the environment is an RP
fixture environment, as denoted by the added environmental variable.

Signed-off-by: Michael Boquard <michael@redpanda.com>