net/probes: Refactor TLS probe for fail safety
Previously, on certificate load, expiry times would be overwritten
iff the expiry of the new cert was sooner than the old stored expiry.
This is a bug. Each time credentials are reloaded, the recorded expiry
should always reflect the soonest expiring cert in the _current_
credential set, irrespective of whatever the probe stored previously.

To that end, this commit refactors tls_certificate_probe such that:
- The default state is "no cert"/"no CA"
- The default expiry time for a cert/CA is time_point::min
- Certificate state is cleared immediately and unconditionally
  at load time.

With these changes, probe state when `!_cert_loaded` can be considered
"junk" but should produce correct semantics for all metrics.