storage: add assertion to supply group id
Offset translator state is stored in the kvstore and keyed by group id.
A default argument is provided for logs' group ids, which makes this
somewhat a footgun. To mitigate, this adds a check that when a set of
types are provided to be translated, a valid group id is also passed.

Note that when there are no translated types, offset translation doesn't
occur and this shouldn't matter.