c/hm: use columnar report in health rpc
Added parameter allowing requester to control the format in which a
health report is going to be returned. The default is to use the old
format. This way updated nodes can still communicate with old nodes as
the not updated node will always default to older health report format
which is supported.

Signed-off-by: Michal Maslanka <michal@redpanda.com>