rptest: split RedpandaServiceK8s into its own file
Currently RedpandaServiceK8s is non-functional: it was intended to be
the base class for tests using plain k8s (i.e., not cloud) but it is
not currently functional.

This change stops short of removing it but moves it to its own file so
that it doesn't interfere with attempts to property type the remaining
redpanda services in redpanda.py.