rptest: move cloud service start() to init
The pattern to use RedpandaServiceCloud was to construct it, then
call start() on it. This is carried over from the pattern for
ducktape Services, but RSC is no longer a service. Having this split
initialization causes a lot of problems for typing, because the type
checker has no guarantee that start() has been called (and in practice
this guarantee is only "by contention"), so it will always complain
about fields which are only set in start().

Just move the body of start() to the constructor and leave start()
empty. Later we can remove start entirely.