storage: add log_reader option to translate offsets
Previously offset translation on the read path was done in the Kafka
layer, with the storage layer providing primitives to help translate.
As we clarify the requirements of the storage log implementation, offset
translation has made its way into the purview of storage.

To allow the storage layer to own offset translation end-to-end on the
read path, and allow for future log implementations that translate
offsets differently (e.g. by storing translated offsets), this adds an
option to the log reader to translate. The corresponding code is removed
from the Kafka layer.