cloud_storage/topic_manifest: implemented manifest_format::serde
introduced topic_manifest_serde proxy class for serialization.
this will help evolve the topic_manifest, if there is the need in the
future.

manifest_format::serde will use the proxy object to perfom
serialization/deserialization.

the default deserialization format remains json, for this commit.

the serialization format is decided with the field _manifest_format.
this is initialized only to topic_manifest::first_version for this
commit.