admin/api-doc: Remove role_member_update def
The motivation for this is that one of the properties specified
in the outward facing API spec has a name collision in generated
code, so the name used in the def didn't match. Since we don't
actually use the generated code, we can remove the definition
altogether and define the structure inline in security.json.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>