sec: s/flat_hash_map/node_hash_map/ in acl_store::remove_bindings
We maintain a running vector of roles principal views to delete from a small
cache maintained by each matched acl_entry_set. The value type is a view to
avoid excessive copying, but that view is always into the principal component
of a key entry (acl_binding) in a flat_hash_map.

This is problematic because any insertion into flat_hash_map might result in
a rehash, invalidating key pointers (including the pointer to key struct member
contained in the principal_view), resulting in use-after-free errors whenever
the following conditions are met:

- More than one binding is deleted from the ACL store
  - i.e. > 1 binding matches the provided filters
- One of those bindings is to a role principal

This commit also adds a test for this (common) situation. I've verified that
*without* this change, the test does indeed crash as expected.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>