raft: error out on empty appends
prior to this change, empty appends were silently passing with incorrect
last offsets set. This is not possible via kafka path because we have
checks that ensure that the records are non empty. Certain internal
paths that use raft APIs directly can potentially try to replicate empty
data, in which case appropriate error is thrown.