rptest: better error messages in Kafka CLI tools
When a KafkaCLI tool fails, we usually get a relative inscrutable
"process returned 1" message in the stack trace. The true error
generally appears on stdout, but you have to dig in the logs to
get that.

This change grabs the last ERROR line in the standard output, which
generally seems to contain the good stuff, and includes it in the
error message.

We also add the name of the operation we were trying to perform to
the error message.