rptest: introduce RedpandaMixedTest
RedpandaMixedTest (RMT) is a base class allowing tests to be written
simultaneously against both vanilla (CDT, docker) and cloud environments
and will run in both of those environments. That is, rather than two
tests deriving from RedpandaTest and RedpandaCloudTest, you could write
one deriving from RMT instead.

The self.redpanda attribute inside RMT has type:

RedpandaService | RedpandaServiceCloud

i.e., it is one or the other. This enables you to call directly the
subset of methods which are common to both or also to narrow the type
using an isinstance check in which case you could do cloud or vanilla
things inside the branches of the if/else.

Fixes redpanda-data/core-internal#199.