admin/security: DELETE roles may also delete ACLs
While deleting a role, also delete ACLs if the `delete_acls` query
param is present and alpha-convertible to bool{true}.

We can do this by submitting a single delete_acls command to the
cluster controller with a "catch-all" filter for any acl entry bound
to the corresponding role principal.

This is NOT atomic with respect to the corresponding role management
command. It is possible that role deletion could succeed while some
or all ACL deletions return an error. In this case, the delete_role
operation will not be undone, but ACLs may be deleted manually on the
client side as needed.