perf: parameterize perf tests with write caching
Existing perf tests are parameterized with write_caching=[on, off]. So
effectively the new runtime of the test suite is doubled. No perf metric
validators were adjusted in this PR, currently the expecation is that
the performance with write caching is no worse than without. The plan is
to adjust them as we get more data in the coming weeks.