rptest: add KafkaClientSecurity abstraction
We manage a lot of kafka client security state (i.e., the auth related
stuff needed to connect to the kafka endpoint) across all our client
wrappers (and "application" wrappers like OMB) with some inconsistency
in the way we apply default configuration from the associated redpanda
service and some missing typing.

This becomes more important as we introduce mixed cloud tests as now
the default security settings are different depending on the underlying
environment: TLS and SASL is always enabled in the cloud, but is off
by default for vanilla RedpandaService.

Introduce KafkaClientSecurity which is a strongly typed class which
encapsulates these connection settings (in their simple non-SASL and
SASL variations, not stuff like oauth currently).

Expose it on RedpandaService and RedpandaServiceCloud. To be consumed
further in subsequent commits in this series.

Issue redpanda-data/core-internal#1199.