rptest: use KafkaClientSecurity in RpkTool
Use KafkaClientSecurity in rpk tool in the idiomatic way we hope to do
for all clients. Basically take the default KafkaClientSecurity object
from the redpanda service and then "merge in" any explicitly specified
username, password, sasl mechanism or tls enabled settings passed into
__init__.

This merging does checks to make sure the overridden values form a
reasonable set and then exposes the associated properties which we
re-expose on RpkTool under the existing names (we no longer store them
as attributes on the class, we just store the merged KCS object).