cloud_storage: Improve throttling metric collection
Previously, the throttling metric was computed before calling
'throttle'. If number of available units was lower than the number of
bytes to read we caclulated deficiency value as a difference between the
two.

In case of concurrency this could lead to incorrect value of the
metric. If two fibers will do this check and find out the available
untis are too low.