Only write experimental_datalake topics to Parquet.
This implements a simple filter based on the topic name. Topics prefixed
with `experimental_datalake_` will be written to Parquet, others will
not.

This introduces some tech debt in the interest of getting a
proof-of-concept working quickly. Eventually we want to make this a
configuration option and not a magic prefix.