Prefix the Parquet path with the topic name.
Before this change, Parquet files had paths like this:
/tmp/parquet_files/0730b620/kafka/experimental_datalake_test_topic/0_22/83-86-372-3-v1.log.3

To read an entire topic would could try
`pandas.read_parquet("/tmp/ parquet_files/")`,
but that would read all data for all topics. Alternatively we could
iterate through the subdirectories,  but that is painful.

After this change, the topic name comes first:
/tmp/parquet_files/experimental_datalake_test_topic/0730b620/kafka/experimental_datalake_test_topic/0_22/83-86-372-3-v1.log.3

So reading a single topic is simply:
`pandas.read_parquet("/tmp/parquet_files/topic_name")`

Tech Debt:
I believe the prefix directories existed to load balance the s3 bucket.
By putting everything for a topic in a single prefix we're breaking the
load balancing in s3. The addition of Iceberg metadata will allow us
to read a single topic regardless of file layout, and we can undo this
change, but this is useful for near-term experimentation.