rm_stm: do not hold producer lock for the duration of the barrier
The lock ensures that the concurrent idempotent requests do not reorder
at the raft layer, i.e if one of the requests within a concurrent batch
of dempotent requests fail to replicate, the subsequent ones fail too.

Upon hitting an OOOSN, we do not need to hold the lock as we are waiting
for the barrier as we know for a fact that request that acquired this
lock is of invalid sequence and does not affect the concurrently running
requests. Holding the lock while the barrier returns just delays legit
concurrent requests.