compression: Introduce lz4 preallocated buffers
An object which holds a pair of preallocated buffers is added. The
buffers are char arrays allocated at startup. State of allocation is
maintained along with the buffers, to ensure correctness when using
these buffers with malloc and free calls. The c api is introduced in the
next commit.

(cherry picked from commit cf44e60391ec6406307b618a2ed4fda0f8605bc7)