k/replicated_partition: do not return offset out of range from follower
When reading from the follower the offset information may not be up to
date. In Redpanda we implemented the behavior that is described in
`KIP-392`. The described mechanism assumes that consumer will reconcile
with the leader after they receive the offset out of range error from
the follower. This is not implemented by the clients hence returning
out of range error resulted in offset reset.

With the fix in this commit we will never return the out of range error
from a node not being a leader. This way consumers will only use leader
to validate its offset and we will avoid spurious resets.

Signed-off-by: Michał Maślanka <michal@redpanda.com>