Upload Parquet files to S3
This adds a new method, `put_parquet_file` that uploads the parquet file
to the appropriate location in S3. New parquet files are uploaded to s3
in a location prefixed by the topic name. This allows them to be read
with a single command like `pandas.read_parquet`.

Tech Debt:
It, like the commits higher in this stack, doesn't handle memory well.
The entire Parquet file is read into an iobuf in memory and then written
to s3. We should be able to pipe the input stream directly to the s3
client.

Also, it does not support retries or anything. I need to understand
the code better to know what happens in the event of a crash. Ideally
we should try uploading any segments that aren't there yet. This may
already be the behavior: since we upload the Parquet file *before* the
segment file, if we fail while uploading parquet the segment won't be
uploaded. The NTP archiver will try again to upload the segment, and
this code will be called again.

If we fail in between writing Parquet and uploading the segment, we
will upload the Parquet file again, which is inefficient, but should not
be incorrect.