rptest: convert schema registry client security
After the updates to KafkaCliClient, it already does the right thing
when SASL is enabled: taking its credentials from the Redpanda
service so there is no need to manually extract and pass the
credentials any more. Convert the schema registry tests to take
advantage by dropping the explicit configuration.

In another place SerdeClient expects a dict of configuration properties
to pass to python scripts, to we use the new to_dict method of
KafkaClientSecurity to get a dict with the expected values.