rptest: update _create_initial_topics
After the updates to KafkaCliClient, it already does the right thing
when SASL is enabled: taking its credentials from the Redpanda
service so there is no need to manually extract and pass the
credentials any more.

This change converts _create_initial_topics in RedpandaTest to take
advantage by dropping the explicit configuration.