serde: support variant
I have an upcoming change to add a variant for a serde transforms
struct, so I am in advance sending out this change to support variant in
serde.

Backwards compat: See the comment on serde::variant, but any
additions/changes to the types in the variant are not backwards or
forward compatible, so we need to handle this in some wrapper struct.

This works by creating a compile time lookup table of variant index ->
a lambda that creates the variant for the type at the given index. Then
we do a runtime dispatch to the static function based on the index of
the function (with a bounds check).

The stuff you can do with templates are 🤯

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>