net/tls_probe: Introduce "trust_file_crc32c" metric
Compute the crc32c checksum on the complete binary contents of the
trust file on certificate creation and reload. Export this value on
the metric iff a trust file is present AND certs are otherwise valid.
Otherwise export 0.

This will be particularly useful for cloud clients who want to verify
the contents of the trust store but don't have direct access to the
filesystem of the host.