c/topics_frontend: break the loop when dispatching to current leader
Changed handling of `finish_partition_update_request` to break the
dispatch loop when one node leader information is stale. Now when node
receives the `finish_partition_update_request` via RPC service it will
not redirect if further to the current controller preventing the request
to be infinitely redirected when node leader information is stale.

Signed-off-by: Michał Maślanka <michal@redpanda.com>