rptest: Increase backoff interval for GCS
The timeing_stress_test performs a lot of TS operations in short time
and in some cases it has to upload manifest frequently (more often than
once per second). The manifest uploads are driven by segment uploads and
retention. Normally, we will try to upload less often (once per 60s or
less) even if we're writing into the partition constantly. But under the
local storage pressure the ntp-archiver is forced to upload the manifest
mroe frequently. The local storage pressure means that the local storage
wants to evict some data but it can't do this unless the manifest is
uploaded and the clean offset is moved forward.

The timing stress test introduces local storage pressure and uploads
manifests frequently. The GCS may throttle us when we're trying to
reupload the manifest faster than once per second. If the initial
backoff has default value of 100ms this is exactly what we will try to
do once some throttling is applied. Redpanda receives SlowDown response
and decides to retry after 100ms, after 200ms, 400ms etc. It never
uploads the manifest and the test fails.

This fix increases the initial backoff to 1000ms if the test is running
on GCS.