model/transform: add offset options
In order to support starting from specific points in time during
deploys, we add a variant that allows starting from the end of the log
or a specific point in time, at which we will resolve to an offset using
a timequery.

Adding a new field to the end of the struct is backwards compatible, so
we are tree to add this new field - the default value will be
latest_offset, which is consistent with existing behavior.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>