transform/processor: support starting from arbitrary times
This commits respects the initial position in the metadata, and means
that when we start processing records we'll start from the deploy time
for newly deployed transforms, while old transforms we will still start
at the end of the log.

This will remove the need for the song and dance around waiting for
transforms to commit before producing records in tests and matches the
expectations and papercuts customers have ran into when trying
transforms.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>