cluster: avoid throwing on duplicate recovery
This adds a conservative check for in-progress recoveries at bootstrap
time: if a recovery is in progress, we will log an error message instead
of throwing.

In practice we don't expect this error code to be hit, but in case of a
bug or somesuch, this seems preferable over throwing and blocking
applying controller messages.

Note that update_in_progress is the only error code to come out of
cluster_recovery_table::apply().