rptest: store started nodes in a set
Prevents bugs caused by the same node appearing multiple times in the
list.

Happens for example when when we kill signal a node and then we start
it. Although we could check manually the presence in the list, it is
more bug prone approach.

There are cases when we want to operate on each started node
concurrently and if we have a node registered multiple times then we run
into very hard to debug race conditions. Example is
RedpandaService.all_up method.