rpk version: lower timeout and retries of api client
This is a simple command that relies on the admin
API and should be fast.
We lower the timeout and the retries to avoid
hanging for clusters that do not have the admin API
avaliable.

It does not work with our cloud cluster, so we
exit early if the profile is from cloud.