cluster: add `self_test_stage` for more verbose self-test reports
Add 'self_test_stage' as an indicator to user of which self-test
routine is currently running. This commit appends the `self_test_stage`
to the `self_test_backend` and `self_test_frontend`.

Currently, a user can request the status of self-test in 'rpk' using
'cluster self-test status'. However, this only indicates whether a
test is running on a node or if the tests are finished, but not which
test is being ran.

After adding bindings on the `rpk` side, this will enable users to see
which self-test is currently running.