rpk: add stage to self-test status
`rpk` will now generate a report of running nodes with node ID,
as well as the current stage of the self test on those running nodes.

This results in more detailed status updates presented to the user
when `cluster self-test status` is ran in `rpk`.

User will now see which test is currently being run on which node,
if a test is running.