transform/rpc: prevent cross shard memory writes
Currently there is code in the transform/rpc layer that does something
along the following:

```
iobuf somedata;
co_await seastar::submit_to(shard_id, [&somedata] {
        somedata = fetch_some_data();
});
```

This executes a function on another CPU and writes memory to a stack
location on the requesting CPU. It's not clear if this is safe to do.

Instead change to use standard seastar mechanisms to pass memory between
cores:

```
iobuf somedata = co_await seastar::submit_to(shard_id, [] {
        return fetch_some_data();
});
```

The actual pieces here are more complicated than the above example, but
this is what motivates the change itself.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>