c/health: deprecate use of node_health_filter when collecting report
When collecting a remote node report a requester is always querying for
the full report without the filter. We may simplify the logic of report
collection by deprecating the filter field from the request. This will
not influence the backward compatibility as the field is not used anyway
and is always empty.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit ab5a19c3b75c83d450fabab2bacfe068ddb52df0)