s3_client: fix lifetime of delete_objects key
An upcoming commit will log the key on error in the s3 client. As is,
delete_objects() would pass in a dummy key to be logged, whose lifetime
would not last to logging time. This commit makes the method co_await to
ensure the variable lasts for the duration of the call.