mvlog: segment reader and readable segment
Introduces a couple of abstractions to be used as a part of the upcoming
new log_reader implementation. These abstractions are:
- readable_segment: A long-lived representation of a segment file, used
  to construct segment_readers.
- segment_reader: A short-lived accessor to the file used to construct
  input_streams. This doesn't enforce any lifecycle relationships
  between segment_readers and input_streams, but it is expected that the
  readers will outlive the streams.

These abstractions are pretty bare bones compared to the existing
storage::segment, storage::segment_reader, and reader handle. There is
some basic accounting done for the readers, but the complexity of a
ref-counted handle doesn't seem critical for now.