verifier/tx: explicit flush before aborting transaction
Kafka client has this optimization where if there is unflushed data
while aborting a transaction, it is silently dropped on the client side
and only the abort request is sent.

With verifiers we want the aborted data in the partition log so the test
code invariants (like no aborted data is read etc) are actually doing
meaningful work.

This patch explicitly flushes the data before aborting the transaction to
ensure data ends up in the log.