consensus: redo background flushing
Original implementation relied on a CV that used a high resolution timer
and with a timer per raft group that proved to expensive at scale. Also,
the flusher fired periodically regardless of whether there is pending
unflushed data from lower acks.

This commit redoes it using a deferred flush that uses a lowres clock
and the the flushes are scheduled only if unflushed data is appended.
Additionally the timers are canceled in favor of background flushes that
are scheduled when flush.bytes is hit. This removes reliance on timers
as much as possible and the timers should never fire in default acks=all
usecases.