k/replicated_partition: fixed querying end offset of an empty log
When querying offset translator we are relying on the property of the
end offset or high watermark being exclusive. Changed the behavior of
log end offset to increment offset in Raft offset space and then
translate to achieve correct behavior when log is ended with a batch
subjected to offset translation.

Fixes: #16612

Signed-off-by: Michał Maślanka <michal@redpanda.com>