cloud_storage: Set EOS when max_offset is reached
The 'partition_record_batch_reader_impl' uses internal
'remote_segment_batch_reader' to fetch data. When the internal reader
reaches the 'max_offset' it returns an empty result. The
'partition_record_batch_reader_impl::do_load_slice' also returns the
empty result in this case. Since the top level reader is invoked by the
'model::record_batch_reader'. It has an internal async loop which is
running until the EOS is reached. If its internal slice is not empty
it's invoking a user provided callback, then it checks if the stream
reached EOS (it exits if this is the case), and then it calls
'load_slice' of the underlying stream (this is the place where
'do_load_slice' method of the 'remote_segment_batch_reader' is invoked).

Since the low level segment reader reached max_offset it will always
return an empty value and the top level record batch reader will loop
forever.

This commit breaks this loop by checking the current offset of the
segment reader if the empty result is returned. If the current offset
overshoot max-offset it resets the low level stream. The top level loop
will exit after that.