Refactor to make arrow_writing_consumer more testable. Add test.
This factors out the IO code that actually writes the Parquet file.
Now the arrow_writing_consumer returns an arrow::Table which can be
inspected by a test. The IO code to write the file remains untested, but
is pretty simple: there is a method that takes a table and writes it to
the given filename.