docs: fix hyperlink to redpanda website in image
From previous README source, it seems that someone was trying to create a
hyperlink to the Redpanda website using our logo. Due to formatting, this
hyperlink did not work, and clicking the image would only take you to the
Github object for the image.

This commit fixes the hyperlink, and clicking the Redpanda logo will now
take you to our platform-capabilities site.