cst: Downgrade error logs to debug
Recently added error logs do not take into account sleep aborted and
other shutdown related errors. Since the logs print exception details
which are already logged down the call chain, and the new logs are for
debugging hang in chunk read path, the new logs are downgraded to debug
level in this change.