archival: Start housekeeping jobs only on a leader
Currently, the housekeeping jobs are enabled in the 'start' method.
First the 'sync' method of the archival metadata stm is called. We don't
check the result of the 'sync' call which is a mistake. Normally, we're
calling 'start' method of the 'ntp_archiver' when the partition is
already a leader so this code works as expected. But if the partition is
moved to another shard we could potentially create and start archiver
when the partition is not a leader yet. In this case the 'sync' call
returns 'nullopt'. The housekeeping jobs are not enabled because
_parent.is_leader() returns 'false' in this case.

Then, when the partition becomes a leader the 'notify_leadership' method
is invoked. This method enables the housekeeping jobs. The problem is
that the 'sync' method of the archival STM may not be called yet. So the
adjacent_segment_merger starts reuploading segments based on stale
manifest.

The fix delays enablement of the housekeeping jobs until the background
loop is started and 'sync' is successfully called. The
'notify_leadership' method can enable or disable the jobs after that.